feat: add Vercel deployment configuration and ignore files
- Created .vercelignore to exclude unnecessary files from deployment.
- Added VERCEL_DEPLOY.md for deployment instructions on Vercel.
- Introduced vercel.json for project configuration and routing.
- Implemented api/index.ts as the entry point for serverless functions, connecting to the database on initialization.

+++ src/services/implementations/FilesService.ts
@@ -32,7 +32,7 @@ export class FilesService implements IFilesService {
         const filePath = path.join(destFolder, nombreUnico);
 
         // Guarda el archivo en el sistema de archivos
-        await fs.promises.writeFile(filePath, file.buffer);
+        await fs.promises.writeFile(filePath, file.buffer as Uint8Array);
 
         // Crea una instancia de Archivo
         const archivo = new Archivo();
